[CCUBE-2183][GZ] refactor cache usage
