+++ tools/storybook-argtypes/services/index.ts
@@ -9,6 +9,6 @@ export { ArgTypesRowBuilder } from "./arg-types-row-builder";
 export { FilePathResolver } from "./file-path-resolver";
 export { JsDocMetadataExtractor } from "./jsdoc-metadata-extractor";
 export { StoryRegistryGenerator } from "./story-registry-generator";
-export { TypeDependencyCache } from "./type-dependency-cache";
+export { TypeDependencyResolver } from "./type-dependency-resolver";
 export { TypeFormattingService } from "./type-formatting-service";
 export { TypeScriptSourceAnalyzer } from "./typescript-source-analyzer";
